requirements.txt
tensorflow_addons==0.16.1
numpy==1.21.5
tensorflow==2.8.0
Pillow==9.0.1
